+++ lib/pages/home_page.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+
+class HomePage extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text("Home"),
+        centerTitle: true,
+        actions: [
+          TextButton(
+            onPressed: () {},
+            child: Text("Log Out",
+              style: TextStyle(
+                color: Colors.white,
+                fontSize: 18,
+              ),
+            ),
+          )
+        ],
+      ),
+    );
+  }
+}
