+++ lib/pages/sign_in_page.dart
@@ -1,10 +1,22 @@
-
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'package:time_tracker/custom_widgets/SignIn_with_email_button.dart';
-import 'package:time_tracker/custom_widgets/facebook_SignIn_button.dart';
-import 'package:time_tracker/custom_widgets/go_anonymous_button.dart';
-import 'package:time_tracker/custom_widgets/google_SignIn_button.dart';
+
 class MySignInPage extends StatelessWidget {
+  const MySignInPage({Key key, @required this.onSignIn}) : super(key: key);
+
+  //A callback when user click on sign in button
+  final void Function(User) onSignIn;
+
+  //A function for Signing In Anonymously
+  Future<void> signInAnonymously() async {
+    try {
+      final userCredential = await FirebaseAuth.instance.signInAnonymously();
+      onSignIn(userCredential.user);
+    } catch (e) {
+      print(e.toString());
+    }
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -21,6 +33,7 @@ class MySignInPage extends StatelessWidget {
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
+
             Text(
               "Sign In",
               style: TextStyle(
@@ -28,26 +41,188 @@ class MySignInPage extends StatelessWidget {
                 fontWeight: FontWeight.w700,
               ),
             ),
+
             SizedBox(
               height: 50,
             ),
-            SignInGoogle(),
+
+            //Sign In using google
+            ElevatedButton(
+              onPressed: () {},
+              style: ButtonStyle(
+                elevation: MaterialStateProperty.all(6),
+                shape: MaterialStateProperty.all(StadiumBorder()),
+                padding: MaterialStateProperty.all(EdgeInsets.all(14)),
+                backgroundColor: MaterialStateProperty.all(Colors.white),
+              ),
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                children: [
+                  SizedBox(
+                    height: 30,
+                    width: 30,
+                    child: Image(
+                      image: AssetImage(
+                        "assets/images/g-logo.png",
+                      ),
+                      fit: BoxFit.scaleDown,
+                    ),
+                  ),
+                  Text(
+                    "Sign In with Google",
+                    style: (TextStyle(
+                      color: Colors.black,
+                    )),
+                  ),
+                  Opacity(
+                    opacity: 0.0,
+                    child: SizedBox(
+                      height: 30,
+                      width: 30,
+                      child: Image(
+                        image: AssetImage(
+                          "assets/images/g-logo.png",
+                        ),
+                        fit: BoxFit.scaleDown,
+                      ),
+                    ),
+                  ),
+                ],
+              ),
+            ),
+
             SizedBox(
               height: 22,
             ),
-            SignInFacebook(),
+
+            //Sign In using facebook
+            ElevatedButton(
+              onPressed: () {},
+              style: ButtonStyle(
+                elevation: MaterialStateProperty.all(6),
+                shape: MaterialStateProperty.all(StadiumBorder()),
+                padding: MaterialStateProperty.all(EdgeInsets.all(14)),
+                backgroundColor: MaterialStateProperty.all(Color(0xff3b5998)),
+              ),
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                children: [
+                  SizedBox(
+                    height: 30,
+                    width: 30,
+                    child: Image(
+                      image: AssetImage(
+                        "assets/images/facbook-logo.jpg",
+                      ),
+                      isAntiAlias: true,
+                      fit: BoxFit.fill,
+                    ),
+                  ),
+                  Text(
+                    "Sign In with Facebook",
+                    style: (TextStyle(
+                      color: Colors.white,
+                    )),
+                  ),
+                  Opacity(
+                    opacity: 0.0,
+                    child: SizedBox(
+                      height: 30,
+                      width: 30,
+                      child: Image(
+                        image: AssetImage(
+                          "assets/images/facbook-logo.jpg",
+                        ),
+                        fit: BoxFit.fill,
+                      ),
+                    ),
+                  ),
+                ],
+              ),
+            ),
+
             SizedBox(
               height: 22,
             ),
-            SignInEmail(),
+
+            //Sign In using Email
+            ElevatedButton(onPressed: (){},
+              style: ButtonStyle(
+                elevation: MaterialStateProperty.all(6),
+                shape: MaterialStateProperty.all(StadiumBorder()),
+                padding: MaterialStateProperty.all(EdgeInsets.all(14)),
+                backgroundColor: MaterialStateProperty.all(Colors.teal[200]),
+              ),
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                children: [
+                  SizedBox(
+                      height: 30,
+                      width: 30,
+                      child:Icon(
+                        Icons.keyboard,
+                        color: Colors.black,
+                      )
+                  ),
+                  Text(
+                    "Sign In with Email",
+                    style: (
+                        TextStyle(
+                          color: Colors.black,
+                        )
+                    ),
+                  ),
+                  Opacity(
+                    opacity: 0.0,
+                    child:  SizedBox(
+                        height: 30,
+                        width: 30,
+                        child:Icon(
+                          Icons.keyboard,
+                          color: Colors.black,
+                        )
+                    ),
+                  ),
+                ],
+              ),
+            ),
+
             SizedBox(
               height: 22,
             ),
+
             Text("or"),
+
             SizedBox(
               height: 22,
             ),
-            SignInAnonymously(),
+
+            //Sign In Anonymously
+            ElevatedButton(
+              onPressed: signInAnonymously,
+              style: ButtonStyle(
+                elevation: MaterialStateProperty.all(6),
+                shape: MaterialStateProperty.all(StadiumBorder()),
+                padding: MaterialStateProperty.all(EdgeInsets.all(14)),
+                backgroundColor: MaterialStateProperty.all(Colors.yellow[50]),
+              ),
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  SizedBox(
+                    height: 30,
+                    child: Center(
+                      child: Text(
+                        "Go Anonymous",
+                        style: (TextStyle(
+                          color: Colors.black,
+                        )),
+                      ),
+                    ),
+                  ),
+                ],
+              ),
+            ),
           ],
         ),
       ),
